refactor(gui): split per-object wiring into MainWindow_Wiring.cpp (#3351 Phase 1d)
Fifth PR of the #3351 monolith-decomposition series. Pure code motion —
no behavior change, no MainWindow.h changes.

Moves the dynamic-object wiring family — the code that runs when a
panadapter, slice, VFO, or DSP widget comes into existence — into the
new MainWindow_Wiring.cpp (2,149 lines):

- wirePanadapter() (1,016 lines — per-pan signal wiring)
- wireVfoWidget() (218)
- wireAetherDspWidget() (67)
- onSliceAdded() / onSliceRemoved() (588 — slice lifecycle)
- panFollowVfo() + revealFrequencyIfNeeded() (pan-follows-tuning policy)

This family is the most RadioSession-adjacent code in the split: when
the session aggregate lands, these methods take a session pointer
instead of reaching for the singular m_radioModel. The new TU's header
comment says so explicitly.

Supporting moves, called out for review:

- Moved with their only callers into the new TU's anon namespace:
  kPanFollowAnimationDurationMs, quantizeIncrementalFollowDelta,
  memoryRevealTargetMatches (+ its tolerance constant), and the four
  pan-follow edge-margin constants.
- Promoted to MainWindowHelpers (now cross-TU):
  - panXpixelsFor / panYpixelsFor / panPixelDimensionsReady (+ their
    five dimension constants, kept private to the helpers TU) — the
    connect-time pan sizing in MainWindow.cpp shares them with the
    moved wiring.
  - xvtrPolicyBandsFrom / xvtrListSummary / xvtrForBandSummary — XVTR
    diagnostics formatters shared by slice-tuning logs (MainWindow.cpp)
    and the moved wiring.

Breadcrumbs at all excision sites.

MainWindow.cpp: 15,239 → 13,095 lines (series: 19,474 → 13,095, −32.8%).
Verified: full build clean, all 33 ctest suites pass, a11y lint clean.

Refs #3351. Principle XI.
