+++ CMakeLists.txt
@@ -638,6 +638,7 @@ set(GUI_SOURCES
     src/gui/MainWindow_Controllers.cpp
     src/gui/MainWindow_Menus.cpp
     src/gui/MainWindow_Shortcuts.cpp
+    src/gui/MainWindow_Wiring.cpp
     src/gui/AgcCalibrationDialog.cpp
     src/gui/AudioDeviceChangeDialog.cpp
     src/gui/ConnectionPanel.cpp
